+++ src/noahtheduke/splint/pattern.clj
@@ -4,60 +4,10 @@
 
 (ns noahtheduke.splint.pattern
   (:require
-    [clojure.string :as str]))
+    [clojure.string :as str]
+    [noahtheduke.splint.utils :refer [drop-quote simple-type]]))
 
 (set! *warn-on-reflection* true)
-
-(defn drop-quote
-  "Convert (quote (a b c)) to (a b c)."
-  [sexp]
-  (if (and (seq? sexp)
-           (= 'quote (first sexp)))
-    (fnext sexp)
-    sexp))
-
-(defprotocol SexpType
-  (simple-type
-    [sexp]
-    "Because Clojure doesn't have this built-in, we must do it the slow way: take
-    an object and return a keyword representing that object:
-
-    ```clojure
-    nil -> :nil
-    true/false -> :boolean
-    \\c -> :char
-    1 -> :number
-    :hello -> :keyword
-    \"hello\" -> :string
-    hello -> :symbol
-    {:a :b} -> :map
-    #{:a :b} -> :set
-    [:a :b] -> :vector
-    (1 2 3) -> :list
-    :else -> (type sexp)
-    ```"))
-
-(extend-protocol SexpType
-  ; literals
-  nil (simple-type [_sexp] :nil)
-  Boolean (simple-type [_sexp] :boolean)
-  Character (simple-type [_sexp] :char)
-  Number (simple-type [_sexp] :number)
-  String (simple-type [_sexp] :string)
-  clojure.lang.Keyword (simple-type [_sexp] :keyword)
-  clojure.lang.Symbol (simple-type [_sexp] :symbol)
-  ; reader macros
-  clojure.lang.IPersistentMap (simple-type [_sexp] :map)
-  clojure.lang.IPersistentSet (simple-type [_sexp] :set)
-  clojure.lang.IPersistentVector (simple-type [_sexp] :vector)
-  clojure.lang.ISeq (simple-type [_sexp] :list)
-  ; else
-  Object (simple-type [sexp] (symbol (pr-str (type sexp)))))
-
-(comment
-  (simple-type {:a 1})
-  (simple-type (Object.))
-  (simple-type `(1 2 3)))
 
 (defn read-dispatch
   "Same as [[simple-type]] except that :symbol and :list provide hints about