+++ src/noahtheduke/splint/rules/style/set_literal_as_fn.clj
@@ -4,9 +4,9 @@
 
 (ns ^:no-doc noahtheduke.splint.rules.style.set-literal-as-fn
   (:require
-    [noahtheduke.splint.pattern :refer [simple-type drop-quote]]
     [noahtheduke.splint.diagnostic :refer [->diagnostic]]
-    [noahtheduke.splint.rules :refer [defrule]]))
+    [noahtheduke.splint.rules :refer [defrule]]
+    [noahtheduke.splint.utils :refer [simple-type drop-quote]]))
 
 (set! *warn-on-reflection* true)