+++ CHANGELOG.md
@@ -34,6 +34,7 @@ The big feature here is adding support to run `splint` without specifying paths.
 - Output formats `simple`, `full,` and `clj-kondo` now print the number of files checked as well: `"Linting took 1ms, checked 3 files, 3 style warnings"`
 - Moved `splint.replace/revert-splint-reader-macros` into `splint.printer` where it belongs.
 - Rely on undefined behavior in `symbol` to correctly print unprintable special characters by converting sexprs to strings and then converting those to symbols.
+- Move `simple-type` and `drop-quote` to `splint.utils`.
 
 ## v1.9.0